Finishing the project

+++ app/Http/Controllers/EventController.php
@@ -71,13 +71,21 @@ class EventController extends Controller
 
         $events = $user->events;
 
-        return view('events.dashboard', ['events' => $events]);
+        $eventsAsParticipant = $user->eventsAsParticipant;
+
+        return view('events.dashboard', ['events' => $events, 'eventsAsParticipant' => $eventsAsParticipant]);
 
     }
 
     public function destroy($id) {
 
+        $user = auth()->user();
+
         $event = Event::findOrFail($id);
+
+        if($user->id != $event->user_id) {
+            return redirect('/dashboard');
+        }
 
         if($event->image) {
             unlink(public_path('img/events/' . $event->image));
@@ -91,7 +99,13 @@ class EventController extends Controller
 
     public function edit($id) {
 
+        $user = auth()->user();
+
         $event = Event::findOrFail($id);
+
+        if($user->id != $event->user_id) {
+            return redirect('/dashboard');
+        }
 
         $event_duration = $this->event_duration($event);
 
@@ -114,6 +128,19 @@ class EventController extends Controller
         $event->update($data);
 
         return redirect('/dashboard')->with('msg', 'Evento editado com sucesso!');
+
+    }
+
+
+    public function joinEvent($id) {
+
+        $user = auth()->user();
+
+        $user->eventsAsParticipant()->attach($id);
+
+        $event = Event::findOrFail($id);
+
+        return redirect('/dashboard')->with('msg', 'Você se inscreveu no evento ' . $event->title);
 
     }
     
@@ -169,16 +196,5 @@ class EventController extends Controller
 
     }
 
-    public function joinEvent($id) {
-
-        $user = auth()->user();
-
-        $user->eventAsParticipant()->attach($id);
-
-        $event = Event::findOrFail($id);
-
-        return redirect('/dashboard')->with('msg', 'Você se inscreveu no evento. ' . $event->title);
-
-    }
 
 }
